requirements.txt
attrs==21.4.0
fasttext==0.9.2
gensim==3.8.3
google-cloud-bigquery==2.31.0
google-cloud-storage==2.0.0
pyahocorasick==1.4.2
pytest==6.2.5
scikit-learn==1.0.2
tqdm==4.62.3